+++ packages/mongo-livedata/mongo_livedata_tests.js
@@ -818,42 +818,6 @@ testAsyncMulti('mongo-livedata - document with a date, ' + idGeneration, [
   }
 ]);
 
-testAsyncMulti('mongo-livedata - transform must return an object, ' + idGeneration, [
-  function (test, expect) {
-    var self = this;
-    var justId = function (doc) {
-      return doc._id;
-    };
-    var idInArray = function (doc) {
-      return [doc._id];
-    };
-    TRANSFORMS["justId"] = justId;
-    var collectionOptions = {
-      idGeneration: idGeneration,
-      transform: justId,
-      transformName: "justId"
-    };
-    var collectionName = Random.id();
-    if (Meteor.isClient) {
-      Meteor.call('createInsecureCollection', collectionName, collectionOptions);
-      Meteor.subscribe('c-' + collectionName);
-    }
-    self.coll = new Meteor.Collection(collectionName, collectionOptions);
-    self.coll.insert({}, expect(function (err, id) {
-      test.isFalse(err);
-      test.isTrue(id);
-      test.throws(function () {
-        self.coll.findOne();
-      });
-      test.throws(function () {
-        self.coll.findOne({}, {transform: idInArray});
-      });
-      // you can still override the transform though.
-      test.equal(self.coll.findOne({}, {transform: null}), {_id: id});
-    }));
-  }
-]);
-
 testAsyncMulti('mongo-livedata - document goes through a transform, ' + idGeneration, [
   function (test, expect) {
     var self = this;
@@ -911,37 +875,6 @@ testAsyncMulti('mongo-livedata - document goes through a transform, ' + idGenera
       test.isFalse(err);
       test.isTrue(id);
       self.id2 = id;
-    }));
-  }
-]);
-
-testAsyncMulti('mongo-livedata - transformed object can\'t have conflicting _id, ' + idGeneration, [
-  function (test, expect) {
-    var self = this;
-    var justId = function (doc) {
-      doc._id = "foo";
-      return doc;
-    };
-    TRANSFORMS["justId"] = justId;
-    var collectionOptions = {
-      idGeneration: idGeneration,
-      transform: justId,
-      transformName: "justId"
-    };
-    var collectionName = Random.id();
-    if (Meteor.isClient) {
-      Meteor.call('createInsecureCollection', collectionName, collectionOptions);
-      Meteor.subscribe('c-' + collectionName);
-    }
-    self.coll = new Meteor.Collection(collectionName, collectionOptions);
-    self.coll.insert({}, expect(function (err, id) {
-      test.isFalse(err);
-      test.isTrue(id);
-      test.throws(function () {
-        self.coll.findOne();
-      });
-      // you can still override the transform though.
-      test.equal(self.coll.findOne({}, {transform: null})._id, id);
     }));
   }
 ]);